Creating Automatic Emissivity Calculator
Did some testing on the emissivity calculation, figuring out that the differences between my method and Thomas' was that Thomas' was using the sin term go correct for the solid angle while I was doing it in my MC propability. I went with Thomas' method and creating a new emissivity calculation script that is made to take in inputs from the command line, and thus it can be used for automation. The script is ./corona_new/emissivity_calc_auto.py.