Adding in FITS file creator for LP model
I fixed the automated LP emissivity calculator to be consistent with
Python3 and I created a new .py script (lp_fits_auto_v1.py) that
creates FITS files for the LP model. It isn't completed, and I will
need to add in another section to append to existing FITS files as
it only works for new FITS files.

Changes to be committed:
	modified:   emissivity_calc_auto.py
	new file:   lp_fits_auto_v1.py