Finishing corona automation, creating utilties library, and starting corona fits file combination script
I have completed the corona automation, having created and successfully ran 5 different .sh files (one
for each Yorp node used). I also had to change the structure of the some of the related Python scripts
slightly in order to scale up to 1000s of parameter combinations. I also started a utilties library
for the text read in function and the functions for the radius of the ISCO and the event horizon. Finally,
I started working on the combine script for the corona fits files that will make them into one giant fits
file.

 Changes to be committed:
	modified:   corona_auto_test.sh
	new file:   corona_full_combine_v1.py
	modified:   create_corona_param_files.py
	modified:   create_corona_params.py
	modified:   lp_fits_auto_v1.py
	modified:   lp_fits_combine_auto_v1.py
	new file:   my_utilities.py
	new file:   yorp11_corona_auto.sh
	renamed:    corona_size_test.sh -> yorp19_corona_auto.sh
	new file:   yorp1_corona_auto.sh
	new file:   yorp2_corona_auto.sh
	new file:   yorp4_corona_auto.sh