Finished first version of full corona FITS combine script and fixed utilities script
I have a seemingly working version of the corona combine script (corona_full_combine_v1.py), though I will have to run tests on it. A by-eye examination
suggests it's good, but I still want to be sure. I also fixed my_utilities.py
to get rid of a bug in read_txt_file which made it produce an error when
dealing with string read-in values.

Changes to be committed:
	modified:   corona_full_combine_v1.py
	modified:   my_utilities.py